pyproject.toml
[tool.poetry]
name = "habits"
version = "0.1.0"
description = ""
authors = ["air <talko.alexander@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
setuptools = "^73.0.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
redis = "^5.0.8"
drf-yasg = "^1.21.7"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
